% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pattern}
\alias{pattern}
\title{Generate a pattern in png format.}
\usage{
pattern(
  type = "bricks",
  density = 8,
  pattern.line.size = 10,
  color = "black",
  background.color = "white",
  pixel = 5,
  res = 30
)
}
\arguments{
\item{type}{pattern types include: 'blank', 'bricks', 'vdashes', 'hdashes', 'crosshatch','dots',
'grid','hlines','nelines','nwlines','vlines','waves','Rsymbol_0' to 'Rsymbol_25', and unicode symbols.}

\item{density}{a numeric number, the density for the lines/dots of a pattern.}

\item{pattern.line.size}{a numeric value, the line size for the lines/dots of a pattern.}

\item{color}{color for the lines/dots of pattern.}

\item{background.color}{color to be filled in the background.}

\item{pixel}{a numeric value, the pixel resolution of the pattern.}

\item{res}{a numeric value, the pixel resolution of the pattern.}
}
\value{
A ggplot object.
}
\description{
The \code{pattern} function is a function for generating a pattern in png format.
}
\details{
\code{pattern} function generates a pattern in png format.
}
\author{
Chunqiao Luo (chunqiaoluo@gmail.com)
}
